Lots of updates!

+++ be/routes/patches.py
@@ -1,57 +1,77 @@
-from datetime import datetime
+import uuid
 
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy import func
-from sqlmodel import select, Session
-from database import get_session
-from models import Patch, Patch
+from sqlmodel import select
+from sqlmodel.ext.asyncio.session import AsyncSession
+
+from auth.user_session_manager import current_active_user
+from database import get_async_session
+from models import Patch, User, Flight
 
 patches_router = APIRouter(tags=["patches"], prefix="/patches")
 
 
 @patches_router.get("")
-def get_patches(session: Session = Depends(get_session)):
-    return session.exec(select(Patch)).all()
+async def get_patches(session: AsyncSession = Depends(get_async_session)):
+    return (await session.exec(select(Patch))).all()
 
 
-@patches_router.get("/{patch_id}")
-def get_patch(user_id: int, flight_id: int, session: Session = Depends(get_session)):
-    patch = session.get(Patch, (user_id, flight_id))
+@patches_router.get("/{id}")
+async def get_patch(id: str, session: AsyncSession = Depends(get_async_session)):
+    patch = await session.get(Patch, id)
     if not patch:
         raise HTTPException(status_code=404, detail="Patch not found")
     return patch
 
 
 @patches_router.post("")
-def create_patch(user_id: int, flight_id: int, patch_number: int | None = None, session: Session = Depends(get_session)):
-    patch = Patch(user_id=user_id, flight_id=flight_id, patch_number=patch_number)
-    # todo: prevent 500 error when adding duplicates
-    # if session.exec(select(func.count((Patch.flight_id, Patch.user_id))).where(Patch.flight_id == flight_id).where(Patch.user_id == user_id)).one():
-    #     raise HTTPException(status_code=400, detail="Patch already exists for this flight + user combo")
+async def create_patch(flight_id: int, patch_number: int | None = None,
+                       session: AsyncSession = Depends(get_async_session), 
+                       user: User = Depends(current_active_user)):
+    if not (await session.exec(select(func.count(Flight.id)).where(Flight.id == flight_id))).one():
+        raise HTTPException(status_code=400, detail="Flight does not exist with that id")
+
+    patch = Patch(user_id=user.id, flight_id=flight_id, patch_number=patch_number)
     session.add(patch)
-    session.commit()
-    session.refresh(patch)
+    await session.commit()
+    await session.refresh(patch)
     return patch
 
 
-@patches_router.put("/{patch_id}")
-def update_patch(user_id: int, flight_id: int, patch_number: int, session: Session = Depends(get_session)):
-    db_patch = session.get(Patch, (user_id, flight_id))
+@patches_router.put("/{id}")
+async def update_patch(id: str, user_id: str | None = None, flight_id: int | None = None, patch_number: int | None = None,
+                       session: AsyncSession = Depends(get_async_session), user: User = Depends(current_active_user)):
+    db_patch = await session.get(Patch, id)
     if not db_patch:
         raise HTTPException(status_code=404, detail="Patch not found")
+    if not user.is_superuser and db_patch.user_id != user.user_id:
+        raise HTTPException(status_code=403, detail="You may only modify your own patches")
 
-    db_patch.patch_number = patch_number
+    if user_id is not None:
+        if not (await session.exec(select(func.count(User.id)).where(User.id == user_id))).one():
+            raise HTTPException(status_code=400, detail="User does not exist with that id")
+        db_patch.user_id = user_id
+    if flight_id is not None:
+        if not (await session.exec(select(func.count(Flight.id)).where(Flight.id == flight_id))).one():
+            raise HTTPException(status_code=400, detail="Flight does not exist with that id")
+        db_patch.flight_id = flight_id
+    if patch_number is not None:
+        db_patch.patch_number = patch_number
     session.add(db_patch)
-    session.commit()
-    session.refresh(db_patch)
+    await session.commit()
+    await session.refresh(db_patch)
     return db_patch
 
 
-@patches_router.delete("/{patch_id}")
-def delete_patch(user_id: int, flight_id: int, session: Session = Depends(get_session)):
-    patch = session.get(Patch, (user_id, flight_id))
+@patches_router.delete("/{id}")
+async def delete_patch(id: str, session: AsyncSession = Depends(get_async_session), user: User = Depends(current_active_user)):
+    patch = await session.get(Patch, id)
     if not patch:
         raise HTTPException(status_code=404, detail="Patch not found")
-    session.delete(patch)
-    session.commit()
+    if not user.is_superuser and patch.user_id != user.id:
+        raise HTTPException(status_code=403, detail="You may only delete your own patches")
+
+    await session.delete(patch)
+    await session.commit()
     return {"message": "Patch deleted successfully"}
